pyproject.toml
[tool.poetry]
name = "bike-rental"
version = "0.1.0"
description = ""
authors = ["AleksandrKosmylev <alex.kosmylev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
